DPL: move ROOT file reading completely to the plugin
This way we can have the plugin depend on JAlien and not
require everyone does.
