+++ Framework/Core/src/AODReaderHelpers.cxx
@@ -175,180 +175,4 @@ AlgorithmSpec AODReaderHelpers::aodSpawnerCallback(std::vector<InputSpec> reques
   }};
 }
 
-AlgorithmSpec AODReaderHelpers::rootFileReaderCallback()
-{
-  auto callback = AlgorithmSpec{adaptStateful([](ConfigParamRegistry const& options,
-                                                 DeviceSpec const& spec,
-                                                 Monitoring& monitoring) {
-    monitoring.send(Metric{(uint64_t)0, "arrow-bytes-created"}.addTag(Key::Subsystem, monitoring::tags::Value::DPL));
-    monitoring.send(Metric{(uint64_t)0, "arrow-messages-created"}.addTag(Key::Subsystem, monitoring::tags::Value::DPL));
-    monitoring.send(Metric{(uint64_t)0, "arrow-bytes-destroyed"}.addTag(Key::Subsystem, monitoring::tags::Value::DPL));
-    monitoring.send(Metric{(uint64_t)0, "arrow-messages-destroyed"}.addTag(Key::Subsystem, monitoring::tags::Value::DPL));
-    monitoring.flushBuffer();
-
-    if (!options.isSet("aod-file")) {
-      LOGP(ERROR, "No input file defined!");
-      throw std::runtime_error("Processing is stopped!");
-    }
-
-    auto filename = options.get<std::string>("aod-file");
-
-    // create a DataInputDirector
-    auto didir = std::make_shared<DataInputDirector>(filename);
-    if (options.isSet("aod-reader-json")) {
-      auto jsonFile = options.get<std::string>("aod-reader-json");
-      if (!didir->readJson(jsonFile)) {
-        LOGP(ERROR, "Check the JSON document! Can not be properly parsed!");
-      }
-    }
-
-    // get the run time watchdog
-    auto* watchdog = new RuntimeWatchdog(options.get<int64_t>("time-limit"));
-
-    // selected the TFN input and
-    // create list of requested tables
-    header::DataHeader TFNumberHeader;
-    std::vector<OutputRoute> requestedTables;
-    std::vector<OutputRoute> routes(spec.outputs);
-    for (auto route : routes) {
-      if (DataSpecUtils::partialMatch(route.matcher, header::DataOrigin("TFN"))) {
-        auto concrete = DataSpecUtils::asConcreteDataMatcher(route.matcher);
-        TFNumberHeader = header::DataHeader(concrete.description, concrete.origin, concrete.subSpec);
-      } else {
-        requestedTables.emplace_back(route);
-      }
-    }
-
-    auto fileCounter = std::make_shared<int>(0);
-    auto numTF = std::make_shared<int>(-1);
-    return adaptStateless([TFNumberHeader,
-                           requestedTables,
-                           fileCounter,
-                           numTF,
-                           watchdog,
-                           didir](Monitoring& monitoring, DataAllocator& outputs, ControlService& control, DeviceSpec const& device) {
-      // check if RuntimeLimit is reached
-      if (!watchdog->update()) {
-        LOGP(INFO, "Run time exceeds run time limit of {} seconds!", watchdog->runTimeLimit);
-        LOGP(INFO, "Stopping reader {} after time frame {}.", device.inputTimesliceId, watchdog->numberTimeFrames - 1);
-        monitoring.flushBuffer();
-        didir->closeInputFiles();
-        control.endOfStream();
-        control.readyToQuit(QuitRequest::Me);
-        return;
-      }
-
-      // Each parallel reader device.inputTimesliceId reads the files fileCounter*device.maxInputTimeslices+device.inputTimesliceId
-      // the TF to read is numTF
-      assert(device.inputTimesliceId < device.maxInputTimeslices);
-      uint64_t timeFrameNumber = 0;
-      int fcnt = (*fileCounter * device.maxInputTimeslices) + device.inputTimesliceId;
-      int ntf = *numTF + 1;
-      monitoring.send(Metric{(uint64_t)ntf, "tf-sent"}.addTag(Key::Subsystem, monitoring::tags::Value::DPL));
-      static int currentFileCounter = -1;
-      static int filesProcessed = 0;
-      if (currentFileCounter != *fileCounter) {
-        currentFileCounter = *fileCounter;
-        monitoring.send(Metric{(uint64_t)++filesProcessed, "files-opened"}.addTag(Key::Subsystem, monitoring::tags::Value::DPL));
-      }
-
-      // loop over requested tables
-      TTree* tr = nullptr;
-      bool first = true;
-      static size_t totalSizeUncompressed = 0;
-      static size_t totalSizeCompressed = 0;
-      static size_t totalReadCalls = 0;
-
-      for (auto route : requestedTables) {
-
-        // create header
-        auto concrete = DataSpecUtils::asConcreteDataMatcher(route.matcher);
-        auto dh = header::DataHeader(concrete.description, concrete.origin, concrete.subSpec);
-
-        // create a TreeToTable object
-        auto info = didir->getFileFolder(dh, fcnt, ntf);
-        size_t before = 0;
-        if (info.file) {
-          info.file->GetReadCalls();
-        }
-        tr = didir->getDataTree(dh, fcnt, ntf);
-        if (!tr) {
-          if (first) {
-            // check if there is a next file to read
-            fcnt += device.maxInputTimeslices;
-            if (didir->atEnd(fcnt)) {
-              LOGP(INFO, "No input files left to read for reader {}!", device.inputTimesliceId);
-              didir->closeInputFiles();
-              control.endOfStream();
-              control.readyToQuit(QuitRequest::Me);
-              return;
-            }
-            // get first folder of next file
-            ntf = 0;
-            tr = didir->getDataTree(dh, fcnt, ntf);
-            if (!tr) {
-              LOGP(FATAL, "Can not retrieve tree for table {}: fileCounter {}, timeFrame {}", concrete.origin, fcnt, ntf);
-              throw std::runtime_error("Processing is stopped!");
-            }
-          } else {
-            LOGP(FATAL, "Can not retrieve tree for table {}: fileCounter {}, timeFrame {}", concrete.origin, fcnt, ntf);
-            throw std::runtime_error("Processing is stopped!");
-          }
-        }
-        TTreePerfStats ps("ioperf", tr);
-
-        if (first) {
-          timeFrameNumber = didir->getTimeFrameNumber(dh, fcnt, ntf);
-          auto o = Output(TFNumberHeader);
-          outputs.make<uint64_t>(o) = timeFrameNumber;
-        }
-
-        // create table output
-        auto o = Output(dh);
-        auto& t2t = outputs.make<TreeToTable>(o);
-
-        // add branches to read
-        // fill the table
-
-        auto colnames = aod::datamodel::getColumnNames(dh);
-        if (colnames.size() == 0) {
-          totalSizeCompressed += tr->GetZipBytes();
-          totalSizeUncompressed += tr->GetTotBytes();
-          t2t.addAllColumns(tr);
-        } else {
-          for (auto& colname : colnames) {
-            TBranch* branch = tr->GetBranch(colname.c_str());
-            totalSizeCompressed += branch->GetZipBytes("*");
-            totalSizeUncompressed += branch->GetTotBytes("*");
-            t2t.addColumn(colname.c_str());
-          }
-        }
-        t2t.fill(tr);
-        if (info.file) {
-          totalReadCalls += info.file->GetReadCalls() - before;
-          static std::string currentFileRead = "";
-          std::string nextFileRead = info.file->GetPath();
-          if (currentFileRead != nextFileRead) {
-            currentFileRead = nextFileRead;
-            monitoring.send(Metric{currentFileRead, "aod-file-read-path"}.addTag(Key::Subsystem, monitoring::tags::Value::DPL));
-          }
-        }
-        monitoring.send(Metric{(double)ps.GetReadCalls(), "aod-tree-read-calls"}.addTag(Key::Subsystem, monitoring::tags::Value::DPL));
-        delete tr;
-
-        first = false;
-      }
-      monitoring.send(Metric{(uint64_t)totalSizeUncompressed / 1000, "aod-bytes-read-uncompressed"}.addTag(Key::Subsystem, monitoring::tags::Value::DPL));
-      monitoring.send(Metric{(uint64_t)totalSizeCompressed / 1000, "aod-bytes-read-compressed"}.addTag(Key::Subsystem, monitoring::tags::Value::DPL));
-      monitoring.send(Metric{(uint64_t)totalReadCalls, "aod-total-read-calls"}.addTag(Key::Subsystem, monitoring::tags::Value::DPL));
-
-      // save file number and time frame
-      *fileCounter = (fcnt - device.inputTimesliceId) / device.maxInputTimeslices;
-      *numTF = ntf;
-    });
-  })};
-
-  return callback;
-}
-
 } // namespace o2::framework::readers
